gayphxmusic/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
minio==7.2.0
boto3==1.34.0
aiofiles==23.2.1
httpx==0.25.2
jinja2==3.1.2
email-validator==2.1.0
reportlab==4.0.7
pillow==10.1.0
ffmpeg-python==0.2.0
pyloudnorm==0.1.1
numpy==1.24.4
scipy==1.11.4
python-dotenv==1.0.0
mutagen==1.47.0
